Separated Post.js into 3 parts: Post.js, PostButton.js, NewCommentInput.js.
